New command decorator
Now, each command is replaced by a closure that checks for varargs
annotations when called. If it is present, the annotation for the varargs is
called with the command's arguments, and the result becomes the new varargs.
This is intended to make a common pattern, looking up the args in the context,
easier (it is now automatic).